added bartok to player branch in preparation for move to master
