--- Util/__init__.py
@@ -1,2 +1,0 @@
-from Util.Deck import Deck, Card, Value, Suit
-from Util.Player import Player
