+++ src/Bartok/__init__.py
@@ -1,0 +1,1 @@
+from src.Bartok.Bartok import Bartok
